refactored and modularized review cards, styles for business page, likes/dislikes/, edit review form
