+++ react-app/src/components/EditReviewForm.js
@@ -1,13 +1,6 @@
 import React from 'react'
 import {
   TextField,
-  Button,
-  Typography,
-  Card,
-  CardActions,
-  CardContent,
-  Grid,
-  Snackbar,
   FormControl,
   Select,
   InputLabel,
@@ -23,7 +16,6 @@ export default function EditReviewForm({
   setTitle,
   setBody,
   setRating}) {
-
 
   const updateTitle = e => {
     setTitle(e.target.value)
